Did some basic weight stuff. Needs to be tweaked later.

+++ EarthWithMagicAPI/API/Interfaces/Items/IItem.cs
@@ -9,7 +9,7 @@ namespace EarthWithMagicAPI.API.Interfaces.Items
         /// <summary>
         /// The weight of the item.
         /// </summary>
-        int Weight { get; }
+        double Weight { get; }
 
         /// <summary>
         /// Holds whether or not the item is equipped.
